Move validated query to models
